+++ extensions/omni_kube/src/api.sql
@@ -2,6 +2,7 @@ create function api(path text,
                     server text default coalesce(current_setting('omni_kube.server', true),
                                                  'https://kubernetes.default.svc'),
                     cacert text default current_setting('omni_kube.cacert', true),
+                    clientcert omni_httpc.client_certificate default row (current_setting('omni_kube.clientcert', true), current_setting('omni_kube.client_private_key', true))::omni_httpc.client_certificate,
                     token text default current_setting('omni_kube.token', true),
                     method omni_http.http_method default 'GET',
                     body jsonb default null) returns jsonb
@@ -36,7 +37,8 @@ begin
     from
         omni_httpc.http_execute_with_options(
                 omni_httpc.http_execute_options(allow_self_signed_cert => cacert is null,
-                                                cacerts => case when cacert is null then null else array [cacert] end),
+                                                cacerts => case when cacert is null then null else array [cacert] end,
+                                                clientcert => case when clientcert is null then null else clientcert end),
                 omni_httpc.http_request(
                         format('%1$s%2$s', server, path),
                         method,